+++ lithops/serverless/backends/aws_batch/aws_batch.py
@@ -522,8 +522,15 @@ class AWSBatchBackend:
 
     def invoke(self, runtime_name, runtime_memory, payload):
         total_calls = payload['total_calls']
+        max_workers = payload['max_workers']
         chunksize = payload['chunksize']
+
+        # Make sure only max_workers are started
         total_workers = total_calls // chunksize + (total_calls % chunksize > 0)
+        if max_workers < total_workers:
+            chunksize = total_calls // max_workers + (total_calls % max_workers > 0)
+            total_workers = total_calls // chunksize + (total_calls % chunksize > 0)
+            payload['chunksize'] = chunksize
 
         job_name = '{}_{}'.format(self._format_jobdef_name(runtime_name, runtime_memory), payload['job_key'])