+++ lithops/standalone/backends/ibm_vpc/ibm_vpc.py
@@ -828,6 +828,12 @@ class IBMVPCBackend:
         """
         return self.config['worker_profile_name']
 
+    def get_worker_cpu_count(self):
+        """
+        Returns the number of CPUs in the worker instance type
+        """
+        return int(self.config['worker_profile_name'].split('-')[1].split('x')[0])
+
     def create_worker(self, name):
         """
         Creates a new worker VM instance