+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [AWS EC2] Changed default image name from "lithops-worker-default" to "lithops-ubuntu-jammy-22.04-amd64-server"
 - [IBM VPC] Changed default image name from "lithops-worker-default" to "lithops-ubuntu-22-04-3-minimal-amd64-1"
 - [Serializer] Improve serializer performance when include_modules is set in config
+- [Standalone] Automatically retrieve te CPU_COUNT from the VM in case worker_processes is not set in config
 
 ### Fixed
 - [Job] Fixed max data size in the invocation payload
@@ -27,7 +28,7 @@
 - [localhost] Get the correct docker/podman path for jobs that run in a container
 - [k8s] Limit the size of the "user" label as the maximum allowed is 63 chars
 - [Joblib] Fix shared objects utility when multiple maps run from the same executor
-- [Docs] Fix documentation regarding execution stats. "host_result_query_count" and "host_status_query_count" were swapped.
+
 
 ## [v3.0.1]